Move Map to map.rs
